requirements.txt
supabase
python-dotenv
icecream
imagekitio
opencv-python
-e .
